feat: iPhonePairingView UI 수정
