+++ LivingStory-iOS/LivingStory-iOS/iPad/iPhonePairing/View/SubView/FindYouriPhoneTextView.swift
@@ -1,0 +1,18 @@
+//
+//  FindYouriPhoneTextView.swift
+//  LivingStory-iOS
+//
+//  Created by 문창재 on 7/21/25.
+//
+import SwiftUI
+
+struct FindYouriPhoneTextView: View {
+    var body: some View {
+        HStack{
+            Text("당신의 아이폰 찾기")
+                .font(.system(size: 13))
+            Spacer()
+        }.padding(.top, 48)
+            .padding(.leading, 20)
+    }
+}
